split system tests in two
